% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{check_plot_tab}
\alias{check_plot_tab}
\title{check_plot_tab}
\usage{
check_plot_tab(App_setting, activate)
}
\arguments{
\item{App_setting}{env App_settings}

\item{activate}{tab to activate}
}
\description{
check_plot_tab
}
