% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{checkPlots}
\alias{checkPlots}
\title{checkPlots}
\usage{
checkPlots(App_settings)
}
\arguments{
\item{App_settings}{environment}
}
\value{

}
\description{
function to check if Plot TabItem is already present, and if
    not it renders the Plots tab.
}
