% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Custom_tables.R
\name{Custom_tables}
\alias{Custom_tables}
\title{Custom_tables}
\value{
The return value, if any, from executing the function.
}
\description{
A fct function
container of R6 objects generator
}
\details{
Custom_tables
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metadata}}{list of metadata}

\item{\code{table1}}{table 1}

\item{\code{table2}}{table 2}

\item{\code{table3}}{table 3}

\item{\code{table4}}{table 4}

\item{\code{table5}}{table 5}

\item{\code{cacheKeys}}{table storing the cacheKeys to avoid reloading the same plots
compile}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-compile}{\code{Custom_tables$compile()}}
\item \href{#method-clone}{\code{Custom_tables$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compile"></a>}}
\if{latex}{\out{\hypertarget{method-compile}{}}}
\subsection{Method \code{compile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Custom_tables$compile(env)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{env}}{App_settings object to access the env containing myCleanMice obj}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Custom_tables$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
