% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plotMod_loader.R
\name{download_obj}
\alias{download_obj}
\title{download_obj}
\usage{
download_obj(title, location, format, ...)
}
\arguments{
\item{title}{}

\item{location}{}

\item{format}{}
}
\value{

}
\description{
download_obj
}
