% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_progress_UI.R
\name{insert_box}
\alias{insert_box}
\title{insert_box}
\usage{
insert_box(placeholder)
}
\arguments{
\item{placeholder}{what is the placeholder to use}
}
\value{
a box
}
\description{
Generates a new box after the placeholder
}
