% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_loader.R
\name{load_data}
\alias{load_data}
\title{load_data}
\usage{
load_data(env)
}
\arguments{
\item{env}{App_settings environment}
}
\description{
creates Clean_mouse_data object and stores into myCleanMice list
}
