% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_progress_UI.R
\name{move_tab}
\alias{move_tab}
\title{move_tab}
\usage{
move_tab(session, tabname)
}
\arguments{
\item{session}{the current R session}

\item{tabname}{the tabItem to be activated}
}
\value{
a different tabItem is activated
}
\description{
activate a specific tabItem.
}
\details{
Calling the move_tab function allows to dynamically change the 
    active tab to a desired one. It is used to seamlessly progress through
    the app when the user finishes the tasks in one section of the app.
}
