% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_progress_UI.R
\name{new_experiment}
\alias{new_experiment}
\title{new_experiment}
\usage{
new_experiment(session, env)
}
\arguments{
\item{session}{the current R session}

\item{env}{to access App_settings}
}
\value{
the app is now ready to receive the files for a new experiment
}
\description{
set up the app for a new experiment
}
\details{
change the app appearance from the landing page to the 
    configuration used to start a new experiment. Restores the dashboard
    header, enables the sidebar and moves the active tab to Input.
}
