% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_progress_UI.R
\name{plot_actogram_sup}
\alias{plot_actogram_sup}
\title{#' insert_box
#' @description Generates a new box after the placeholder
#' @param placeholder what is the placeholder to use
#'
#' @return a box 
#' @export
#'
  insert_box <- function(placeholder, title, plotObj, session){
    browser()
    # plotObj <- plotObj
    shiny::insertUI(
      selector = placeholder,
      where = "afterEnd",
      ui = mod_box_plot_ui(NS("box_plot_ui_1"), title)
    )
    mod_box_plot_server("box_plot_ui_1", plotObj)
    # output$plot_holder <- shiny::renderUI({
    #                           ns <- session$ns
    #                           shiny::renderPlot(plotObj)
    #                         })
  }
plot_actogram_sup}
\usage{
plot_actogram_sup(env, plot_type)
}
\arguments{
\item{env}{aa}

\item{plot_type}{ss}
}
\description{
#' insert_box
#' @description Generates a new box after the placeholder
#' @param placeholder what is the placeholder to use
#'
#' @return a box 
#' @export
#'
  insert_box <- function(placeholder, title, plotObj, session){
    browser()
    # plotObj <- plotObj
    shiny::insertUI(
      selector = placeholder,
      where = "afterEnd",
      ui = mod_box_plot_ui(NS("box_plot_ui_1"), title)
    )
    mod_box_plot_server("box_plot_ui_1", plotObj)
    # output$plot_holder <- shiny::renderUI({
    #                           ns <- session$ns
    #                           shiny::renderPlot(plotObj)
    #                         })
  }
plot_actogram_sup
}
