% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_R6_Raw_mouse_data.R
\name{Raw_mouse_data}
\alias{Raw_mouse_data}
\title{Raw_mouse_data 
#'}
\description{
Raw_mouse_data R6 class object
}
\details{
A R6 class object to contain the uploaded files and info extracted
    from the metadata file. performs first round of preprocessing to store
    data in a convenient format to be further processed.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{animal's id}

\item{\code{sex}}{animal's sex}

\item{\code{genotype}}{animal's genotype}

\item{\code{cabinet}}{cabined in which the animal is housed}

\item{\code{lightOn}}{hour when lights are switched on}

\item{\code{startdate}}{starting date of the experiment}

\item{\code{data}}{raw data}

\item{\code{length}}{length of data table}

\item{\code{fct_1}}{additional user defined factor}

\item{\code{fct_2}}{additional user defined factor
                            
initialize}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Raw_mouse_data-new}{\code{Raw_mouse_data$new()}}
\item \href{#method-Raw_mouse_data-add}{\code{Raw_mouse_data$add()}}
\item \href{#method-Raw_mouse_data-clone}{\code{Raw_mouse_data$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Raw_mouse_data-new"></a>}}
\if{latex}{\out{\hypertarget{method-Raw_mouse_data-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Raw_mouse_data$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
initializes some field with the required type

add
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Raw_mouse_data-add"></a>}}
\if{latex}{\out{\hypertarget{method-Raw_mouse_data-add}{}}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Raw_mouse_data$add(datapath, metadatapath, env, skipRows = 4)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datapath}}{path where to find the uploaded data filed}

\item{\code{metadatapath}}{path where to find the metadata file}

\item{\code{env}}{App_settings environment}

\item{\code{skipRows}}{numeric, number of initial rows to skip in the imported files}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Raw_mouse_data-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Raw_mouse_data-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Raw_mouse_data$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
