% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_progress_UI.R
\name{render_tabItem_ui}
\alias{render_tabItem_ui}
\title{render_tabItem_UI}
\usage{
render_tabItem_ui(tabname, text, icon)
}
\arguments{
\item{tabname}{a character string containing the name of the tabItem to
render. It must match one of the menuItemOutput}

\item{text}{a character string containing the text to be displayed in the 
sidebar}

\item{icon}{a character string containing the icon to appear on the left}
}
\value{
A new menuItem is generated in the app sidebar.
}
\description{
function to dynamically render sidebar menu as the user 
    progresses through the experiment
}
\details{
Call to the function render_tabItem_UI() requires the specification
    of params tabname, text, and icon to render a menuItem in the sidebar
    of the dahboardSidebar. Calling of the function can be associated to
    a call to movetab() to make it the active tab.
}
