% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{subset_input_check}
\alias{subset_input_check}
\title{subset_input_check}
\usage{
subset_input_check(idlist)
}
\arguments{
\item{idlist}{list of selected animals for subsetting}
}
\value{
an error notification
}
\description{
subset_input_check
}
