% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_progress_UI.R
\name{update_DS_ui}
\alias{update_DS_ui}
\title{update_DS_ui}
\usage{
update_DS_ui(case, ...)
}
\arguments{
\item{case}{which case to apply}

\item{...}{additional argument to control switch}
}
\value{

}
\description{
function to dynalically update the possible choices in the
    Data Structure tab box 1
}
