% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_loader.R
\name{upload_subsetting}
\alias{upload_subsetting}
\title{upload_subsetting}
\usage{
upload_subsetting(funEnv, session, input_result)
}
\arguments{
\item{funEnv}{connects to App_settings object}

\item{session}{session object}
}
\description{
upload_subsetting
}
\examples{
upload_subsetting(App_settings)
}
