% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_App_settings.R
\name{App_settings}
\alias{App_settings}
\title{App_settings}
\description{
App_settings R6 class object
}
\details{
A R6 class object to contain parameters of the app that can be
    accessible from the global.env, making possible to communicate between 
    different R6 objects
}
\examples{

## ------------------------------------------------
## Method `App_settings$setData`
## ------------------------------------------------

setMeta


## ------------------------------------------------
## Method `App_settings$setMeta`
## ------------------------------------------------

setListMice


## ------------------------------------------------
## Method `App_settings$setListMice`
## ------------------------------------------------

setListMiceFiltered


## ------------------------------------------------
## Method `App_settings$setListMiceFiltered`
## ------------------------------------------------

saveLDparams


## ------------------------------------------------
## Method `App_settings$saveLDparams`
## ------------------------------------------------

setLD


## ------------------------------------------------
## Method `App_settings$setLD`
## ------------------------------------------------

updateTimeRange


## ------------------------------------------------
## Method `App_settings$updateTimeRange`
## ------------------------------------------------

setDiscRow


## ------------------------------------------------
## Method `App_settings$setDiscRow`
## ------------------------------------------------

setTimeDisp


## ------------------------------------------------
## Method `App_settings$setTimeDisp`
## ------------------------------------------------

setExpstart


## ------------------------------------------------
## Method `App_settings$setExpstart`
## ------------------------------------------------

setTimepointDur


## ------------------------------------------------
## Method `App_settings$setTimepointDur`
## ------------------------------------------------

no examples provided
Initialize_DS

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dashboardHeader_appearance}}{TRUE/FALSE to enable/disable dashboard header}

\item{\code{fileType}}{type of files uploaded by user}

\item{\code{dataList}}{is a list of file paths uploaded by the user}

\item{\code{metadata}}{metadata filepath}

\item{\code{listMice}}{list of animals}

\item{\code{discardRow}}{how many rows to discard from each data file}

\item{\code{timeDisplay}}{how to display time}

\item{\code{ExpStart}}{what day and hour does the experiment starts?}

\item{\code{timepointDur}}{what is the sampling frequency?}

\item{\code{subsetting}}{list containing info about data subsetting}

\item{\code{LDcondition}}{list containing info on LD conditions}

\item{\code{plotTab}}{TRUE/FALSE value to know if the plot tab has been activated}

\item{\code{env1}}{contains environment#1}

\item{\code{env2}}{contains environment#2}

\item{\code{env3}}{contains environment#3}

\item{\code{env4}}{contains environment#4}

\item{\code{env_msg}}{contains the environment of Message obj

setData}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-setData}{\code{App_settings$setData()}}
\item \href{#method-setMeta}{\code{App_settings$setMeta()}}
\item \href{#method-setListMice}{\code{App_settings$setListMice()}}
\item \href{#method-setListMiceFiltered}{\code{App_settings$setListMiceFiltered()}}
\item \href{#method-saveLDparams}{\code{App_settings$saveLDparams()}}
\item \href{#method-setLD}{\code{App_settings$setLD()}}
\item \href{#method-updateTimeRange}{\code{App_settings$updateTimeRange()}}
\item \href{#method-setDiscRow}{\code{App_settings$setDiscRow()}}
\item \href{#method-setTimeDisp}{\code{App_settings$setTimeDisp()}}
\item \href{#method-setExpstart}{\code{App_settings$setExpstart()}}
\item \href{#method-setTimepointDur}{\code{App_settings$setTimepointDur()}}
\item \href{#method-initialize_DS}{\code{App_settings$initialize_DS()}}
\item \href{#method-clone}{\code{App_settings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setData"></a>}}
\if{latex}{\out{\hypertarget{method-setData}{}}}
\subsection{Method \code{setData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App_settings$setData(outsideData)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{outsideData}}{list of data files uploaded by the user}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{setMeta

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setMeta"></a>}}
\if{latex}{\out{\hypertarget{method-setMeta}{}}}
\subsection{Method \code{setMeta()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App_settings$setMeta(outsideMeta)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{outsideMeta}}{metadata file uploaded by the user}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{setListMice

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setListMice"></a>}}
\if{latex}{\out{\hypertarget{method-setListMice}{}}}
\subsection{Method \code{setListMice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App_settings$setListMice(env)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{env}}{environment}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{setListMiceFiltered

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setListMiceFiltered"></a>}}
\if{latex}{\out{\hypertarget{method-setListMiceFiltered}{}}}
\subsection{Method \code{setListMiceFiltered()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App_settings$setListMiceFiltered(x, idList, sexList, genList, cabList)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{environment}

\item{\code{idList}}{list containing mice ids}

\item{\code{sexList}}{list containing mice sex}

\item{\code{genList}}{list containing mice genotypes}

\item{\code{cabList}}{list containing cabinet numbers}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{saveLDparams

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveLDparams"></a>}}
\if{latex}{\out{\hypertarget{method-saveLDparams}{}}}
\subsection{Method \code{saveLDparams()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App_settings$saveLDparams(x, light, ddVal, ddStart)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{App_settings environment}

\item{\code{light}}{ligthOn value}

\item{\code{ddVal}}{Y/N value for DD}

\item{\code{ddStart}}{day in which DD starts}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{setLD

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setLD"></a>}}
\if{latex}{\out{\hypertarget{method-setLD}{}}}
\subsection{Method \code{setLD()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App_settings$setLD(x, light, ddVal, ddStart)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{App_settings environment}

\item{\code{light}}{ligthOn value}

\item{\code{ddVal}}{Y/N value for DD}

\item{\code{ddStart}}{day in which DD starts}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{updateTimeRange

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateTimeRange"></a>}}
\if{latex}{\out{\hypertarget{method-updateTimeRange}{}}}
\subsection{Method \code{updateTimeRange()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App_settings$updateTimeRange(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{to be rechecked (numeric)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{setDiscRow

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setDiscRow"></a>}}
\if{latex}{\out{\hypertarget{method-setDiscRow}{}}}
\subsection{Method \code{setDiscRow()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App_settings$setDiscRow(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{number of rows to discard (numeric)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{setTimeDisp

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setTimeDisp"></a>}}
\if{latex}{\out{\hypertarget{method-setTimeDisp}{}}}
\subsection{Method \code{setTimeDisp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App_settings$setTimeDisp(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{to display time in 12/24h format}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{setExpstart

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setExpstart"></a>}}
\if{latex}{\out{\hypertarget{method-setExpstart}{}}}
\subsection{Method \code{setExpstart()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App_settings$setExpstart(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{date value for when the experiment starts}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{setTimepointDur

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setTimepointDur"></a>}}
\if{latex}{\out{\hypertarget{method-setTimepointDur}{}}}
\subsection{Method \code{setTimepointDur()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App_settings$setTimepointDur(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{sampling rate}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{no examples provided
Initialize_DS

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initialize_DS"></a>}}
\if{latex}{\out{\hypertarget{method-initialize_DS}{}}}
\subsection{Method \code{initialize_DS()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App_settings$initialize_DS(server)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{server}}{is it really needed?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App_settings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
