% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Custom_tables.R
\name{Custom_tables}
\alias{Custom_tables}
\title{Custom_tables}
\value{
The return value, if any, from executing the function.
}
\description{
A R6 object that contains methods to analyse the uploaded data.
    The secondary data that are generated are stored inside the object.
}
\details{
Custom_tables

method to generate a metadata table, integrating the metadata file
    uploaded from the user with computing of each raw data file in the 
    dataset.
}
\examples{

## ------------------------------------------------
## Method `Custom_tables$AvgDay`
## ------------------------------------------------

HHActivity(env)
CheckIf


## ------------------------------------------------
## Method `Custom_tables$checkIf`
## ------------------------------------------------

write examples
behavrTable


## ------------------------------------------------
## Method `Custom_tables$behavrTable`
## ------------------------------------------------

write expamples of behavrTable application
dailyAct


## ------------------------------------------------
## Method `Custom_tables$dailyAct`
## ------------------------------------------------

dailyAct(env)
computePer


## ------------------------------------------------
## Method `Custom_tables$computePer`
## ------------------------------------------------

provide examples on how to call the function for the different 
    period analyses available with different params
    
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metadata}}{list of metadata}

\item{\code{table1}}{table 1}

\item{\code{table2}}{table 2}

\item{\code{table3}}{table 3}

\item{\code{table4}}{table 4}

\item{\code{table5}}{table 5}

\item{\code{cacheKeys}}{table storing the cacheKeys to avoid reloading the same plots
compile}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-compile}{\code{Custom_tables$compile()}}
\item \href{#method-AvgDay}{\code{Custom_tables$AvgDay()}}
\item \href{#method-checkIf}{\code{Custom_tables$checkIf()}}
\item \href{#method-behavrTable}{\code{Custom_tables$behavrTable()}}
\item \href{#method-dailyAct}{\code{Custom_tables$dailyAct()}}
\item \href{#method-computePer}{\code{Custom_tables$computePer()}}
\item \href{#method-clone}{\code{Custom_tables$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compile"></a>}}
\if{latex}{\out{\hypertarget{method-compile}{}}}
\subsection{Method \code{compile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Custom_tables$compile(env)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{env}}{App_settings object to access the env containing myCleanMice obj}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Metadata table to be stored inside the Custom_tables R6 object.

HHActivity
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AvgDay"></a>}}
\if{latex}{\out{\hypertarget{method-AvgDay}{}}}
\subsection{Method \code{AvgDay()}}{
function to create tibble with mouse activity daily grouped in 
    15 minutes bins
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Custom_tables$AvgDay(env, per_len)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{env}}{App_settings environment}

\item{\code{per_len}}{customizable period length value, used to normalise Circadian day}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A table with the sum
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{HHActivity(env)
CheckIf

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkIf"></a>}}
\if{latex}{\out{\hypertarget{method-checkIf}{}}}
\subsection{Method \code{checkIf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Custom_tables$checkIf(funEnv, subsetPlot)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{funEnv}}{env containing App_settings}

\item{\code{subsetPlot}}{Yes/No value}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{write examples
behavrTable

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-behavrTable"></a>}}
\if{latex}{\out{\hypertarget{method-behavrTable}{}}}
\subsection{Method \code{behavrTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Custom_tables$behavrTable(x, subsetVal)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{App_settings environment}

\item{\code{subsetVal}}{Yes/No value if to subset dataset or not}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{write expamples of behavrTable application
dailyAct

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dailyAct"></a>}}
\if{latex}{\out{\hypertarget{method-dailyAct}{}}}
\subsection{Method \code{dailyAct()}}{
function to generate the sum of the activity for each day and
    store it on a table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Custom_tables$dailyAct(env)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{env}}{environment containing myCleanMice object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{dailyAct(env)
computePer

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-computePer"></a>}}
\if{latex}{\out{\hypertarget{method-computePer}{}}}
\subsection{Method \code{computePer()}}{
function to analyse data generated with behavrTable() to 
    extract period lenght in data if present.
    Describe Analysis methods and link to papers that compare the different
    methods
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Custom_tables$computePer(method, periodRange, funenv)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{method to choose. Available are: "chi_sq_periodogram", 
"ac_periodogram", "ls_periodogram".}

\item{\code{periodRange}}{numeric,}

\item{\code{funenv}}{Not used?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{provide examples on how to call the function for the different 
    period analyses available with different params
    
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Custom_tables$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
