% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_R6_Custom_tables.R
\name{Custom_tables}
\alias{Custom_tables}
\title{Custom_tables}
\value{
The return value, if any, from executing the function.
}
\description{
A R6 object that contains methods to analyse the uploaded data.
    The secondary data that are generated are stored inside the object.
}
\details{
Custom_tables

method to generate a metadata table, integrating the metadata file
    uploaded from the user with computing of each raw data file in the 
    dataset.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metadata}}{list of metadata}

\item{\code{locomotor_act}}{list containing tables related to actograms}

\item{\code{daily_act}}{list containing tables related to sum of daily activity}

\item{\code{average_day}}{list containing tables related to average circadian day}

\item{\code{periodograms}}{list containing tables related to periodograms}

\item{\code{cacheKeys}}{table storing the cacheKeys to avoid reloading the same plots
compile}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-compile}{\code{Custom_tables$compile()}}
\item \href{#method-AvgDay}{\code{Custom_tables$AvgDay()}}
\item \href{#method-checkIf}{\code{Custom_tables$checkIf()}}
\item \href{#method-behavrTable}{\code{Custom_tables$behavrTable()}}
\item \href{#method-dailyAct}{\code{Custom_tables$dailyAct()}}
\item \href{#method-computePer}{\code{Custom_tables$computePer()}}
\item \href{#method-clone}{\code{Custom_tables$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compile"></a>}}
\if{latex}{\out{\hypertarget{method-compile}{}}}
\subsection{Method \code{compile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Custom_tables$compile(env)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{env}}{App_settings object to access the env containing myCleanMice obj}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Metadata table to be stored inside the Custom_tables R6 object.

AvgDay
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AvgDay"></a>}}
\if{latex}{\out{\hypertarget{method-AvgDay}{}}}
\subsection{Method \code{AvgDay()}}{
function to create tibble with mouse activity daily grouped in 
    15 minutes bins
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Custom_tables$AvgDay(env, per_len, subsetVal)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{env}}{App_settings environment}

\item{\code{per_len}}{customizable period length value, used to normalise Circadian day}

\item{\code{subsetVal}}{Yes/No value}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A table with the sum 

CheckIf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkIf"></a>}}
\if{latex}{\out{\hypertarget{method-checkIf}{}}}
\subsection{Method \code{checkIf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Custom_tables$checkIf(env, subsetPlot)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{env}}{env containing App_settings}

\item{\code{subsetPlot}}{Yes/No value

behavrTable}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-behavrTable"></a>}}
\if{latex}{\out{\hypertarget{method-behavrTable}{}}}
\subsection{Method \code{behavrTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Custom_tables$behavrTable(env, subsetVal)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{env}}{App_settings environment}

\item{\code{subsetVal}}{Yes/No value if to subset dataset or not

dailyAct}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dailyAct"></a>}}
\if{latex}{\out{\hypertarget{method-dailyAct}{}}}
\subsection{Method \code{dailyAct()}}{
function to generate the sum of the activity for each day and
    store it on a table.

computePer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Custom_tables$dailyAct(env, subsetVal)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{env}}{environment containing myCleanMice object}

\item{\code{subsetVal}}{Yes/No value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-computePer"></a>}}
\if{latex}{\out{\hypertarget{method-computePer}{}}}
\subsection{Method \code{computePer()}}{
function to analyse data generated with behavrTable() to 
    extract period lenght in data if present.
    Describe Analysis methods and link to papers that compare the different
    methods
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Custom_tables$computePer(method, periodRange, env, subsetVal)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{method to choose. Available are: "chi_sq_periodogram", 
"ac_periodogram", "ls_periodogram".}

\item{\code{periodRange}}{numeric,}

\item{\code{env}}{App_settings environment}

\item{\code{subsetVal}}{Yes/No value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Custom_tables$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
