% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plotMod_loader.R
\name{download_obj}
\alias{download_obj}
\title{download_obj}
\usage{
download_obj(title, location, format, ...)
}
\arguments{
\item{title}{Title of the document to download}

\item{location}{path to object location}

\item{format}{extension of the file to output ("csv", "png", "2.csv")}
}
\value{
a file to be saved locally
}
\description{
Function to download data tables and plots
}
\examples{
download_obj(title = "Plot_1", location = Histo_plot, format = "csv")
}
