% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_progress_UI.R
\name{update_DS_ui}
\alias{update_DS_ui}
\title{update_DS_ui}
\usage{
update_DS_ui(session, element, case)
}
\arguments{
\item{session}{env}

\item{element}{subsetting or serietype}

\item{case}{which case to apply}
}
\value{

}
\description{
function to dynalically update the possible choices in the
    Data Structure tab box 1
}
